# apps/pitch-rosetta/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pitch Rosetta API",
    "description": "HTTP API for generating landing pages from business/dental data. Use Bearer token (DEMO_GENERATOR_API_KEY or PITCH_ROSETTA_API_KEY) for protected endpoints.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" }
  ],
  "tags": [
    { "name": "Health", "description": "Readiness and liveness" },
    { "name": "Test", "description": "Development and testing" },
    { "name": "Generate", "description": "AI-powered full generation (sync/async)" },
    { "name": "Dental", "description": "Dental template render (sync/async)" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "getHealth",
        "tags": ["Health"],
        "summary": "Health check",
        "description": "Readiness probe.",
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "service": { "type": "string", "example": "pitch-rosetta-agent" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/render-test": {
      "get": {
        "operationId": "getRenderTest",
        "tags": ["Test"],
        "summary": "Render test (dental HTML)",
        "description": "Returns rendered dental HTML. Optional: json (fixture filename), style (dental-v1..dental-v6). Omit style for a random style guide. Optional layout=lead-rosetta for the Lead Rosetta single layout (only when style is omitted).",
        "security": [],
        "parameters": [
          { "name": "json", "in": "query", "description": "index.json | index-dental-riverside.json | index-dental-downtown.json", "schema": { "type": "string" } },
          { "name": "style", "in": "query", "description": "dental-v1 | dental-v2 | dental-v3 | dental-v4 | dental-v6", "schema": { "type": "string" } },
          { "name": "layout", "in": "query", "description": "lead-rosetta — use Lead Rosetta layout instead of style guides (ignored if style is set)", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "HTML", "content": { "text/html": { "schema": { "type": "string" } } } },
          "503": { "description": "No dental JSON found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/generate": {
      "post": {
        "operationId": "postGenerate",
        "tags": ["Generate"],
        "summary": "Generate (sync) — DISABLED",
        "description": "DISABLED. This endpoint returns 503. Use POST /api/dental or POST /api/dental-async instead. Previously: generated landing page from business data (id, publicUrl, plan, branding, no HTML).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Business data (index.json shape)"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "publicUrl": { "type": "string", "nullable": true },
                    "plan": { "type": "object" },
                    "branding": { "type": "object" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid body", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Generation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "Endpoint disabled", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/generate-async": {
      "post": {
        "operationId": "postGenerateAsync",
        "tags": ["Generate"],
        "summary": "Generate (async + callback) — DISABLED",
        "description": "DISABLED. This endpoint returns 503. Use POST /api/dental-async instead. Previously: accepted job, returned 202, ran generate in background and POSTed result to callbackUrl.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["jobId", "prospectId", "callbackUrl", "callbackToken"],
                "properties": {
                  "jobId": { "type": "string" },
                  "prospectId": { "type": "string" },
                  "userId": { "type": "string" },
                  "callbackUrl": { "type": "string", "format": "uri" },
                  "callbackToken": { "type": "string" }
                },
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "status": { "type": "string", "example": "accepted" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid body or callback URL", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "Endpoint disabled", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/dental": {
      "post": {
        "operationId": "postDental",
        "tags": ["Dental"],
        "summary": "Dental render (sync)",
        "description": "Renders dental HTML from JSON body. Picks a random dental style guide (dental-v1..dental-v6) unless optional styleId or dentalStyleId is set, or leadRosettaLayout is true.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Dental business data (index.json shape). Optional render hints: styleId, dentalStyleId (dental-v1..dental-v6), leadRosettaLayout (boolean). These are stripped before render."
              }
            }
          }
        },
        "responses": {
          "200": { "description": "HTML", "content": { "text/html": { "schema": { "type": "string" } } } },
          "400": { "description": "Invalid body", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Render failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/dental-async": {
      "post": {
        "operationId": "postDentalAsync",
        "tags": ["Dental"],
        "summary": "Dental render (async + callback)",
        "description": "Accepts job, returns 202. Renders dental HTML (random style guide unless styleId/dentalStyleId/leadRosettaLayout in body) and POSTs { jobId, prospectId, userId?, html } or { ..., error } to callbackUrl.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["jobId", "prospectId", "callbackUrl", "callbackToken"],
                "properties": {
                  "jobId": { "type": "string" },
                  "prospectId": { "type": "string" },
                  "userId": { "type": "string" },
                  "callbackUrl": { "type": "string", "format": "uri" },
                  "callbackToken": { "type": "string" }
                },
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "status": { "type": "string", "example": "accepted" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid body or callback URL", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "DEMO_GENERATOR_API_KEY or PITCH_ROSETTA_API_KEY"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" }
        }
      }
    }
  }
}